# doc/openapi.yaml
openapi: "3.0.2"
info:
  title: Friend Finder
  version: '1.0'
  description: "L'API de FriendFinder est une API REST. Elle a des URL orientées ressources, elle renvoie des données codées en JSON, elle utilise les codes de réponses standards HTTP, et une authentification."

  #termsOfService: "https://openweathermap.org/terms"
  license:
    name: "Repo Git"
    url: "https://github.com/EnseirbTelecom/pg219-2020-dark-heroes/tree/master/doc"
servers:
  - url: https://localhost:3000
security:
  - friendFinder_auth: []
...
components:
  securitySchemes:
    friendFinder_auth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
      description: "L'utilisateur doit s'identifier pour accéder aux requêtes."
  schemas:
    Hist_position:
      type: object
      properties:
        hist_lat:
          type: string
        hist_long:
          type: string
        hist_date:
          type: string
          format: date
        message:
          type: string
...
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        first_name:
          type: string
        last_name:
          type: string
        pseudo:
          type: string
        birthday:
          type: string
          format: date
        gender:
          type: string
        current_lat:
          type: string
        current_long:
          type: string
        activation_date:
          type: string
          format: date
        duration:
          type: number
          minimum: 0
        hist_positions:
          $ref: "#/components/schemas/Hist_position"
        friends:
          type: array
          items:
            type: string
        friend_request:
          type: array
          items:
            type: string
        friend_pending:
          type: array
          items:
            type: string
paths:
...
  /friends:
    get:
      tags:
        - Amis
      summary: "Accéder à la liste des amis"
      operationId: "getFriends"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  friendslist:
                    type: array
                    items:
                      type: string
                      example: Antoine Dubois, Sophie Martin, Fred Lanvin
                  status:
                    type: integer
                    example: 200
        "404":
          description: "Erreur"
        "403":
          description: "Erreur"
  /friendsrequests:
    get:
      tags:
        - Amis
      summary: "Accéder à la liste des demandes d'amis"
      operationId: "getFriendrequests"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  friendrequests:
                    type: array
                    items:
                      type: string
                      example: Antoine Dubois, Sophie Martin, Fred Lanvin
                  status:
                    type: integer
                    example: 200
        "404":
          description: "Erreur"
        "403":
          description: "Erreur"
  /delFriend:
    delete:
      tags:
        - Amis
      summary: "Supprimer un ami de la liste"
      operationId: "deletefriend"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "Jean Dupont has been removed."
                  status:
                    type: integer
                    example: 200
        "404":
          description: "Erreur"
          content:
            json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Jean Dupont doesn't exist."
                  status:
                    type: integer
                    example: 404
        "403":
          description: "Erreur"
  /addFriend:
    post:
      tags:
        - Amis
      summary: "Ajouter un ami"
      operationId: "addFriend"
      requestBody:
        description: "Structure User de l'ami à ajouter"
        required: true
        content:
          json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "Friend request for Jean Dupont sent."
                  status:
                    type: integer
                    example: 200

        "400":
          description: "L'ami est déjà présent dans la liste"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Jean Dupont is already your friend."
                  status:
                    type: integer
                    example: 400
        "404":
          description: "L'utilisateur n'existe pas."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Jean Dupont doesn't exist."
                  status:
                    type: integer
                    example: 404
        "403":
          description: "Erreur"
  /friendRequestReply:
    post:
      tags:
        - Amis
      summary: "Répondre à une demande d'ami"
      operationId: "friendRequestReply"
      requestBody:
        description: "JSON contenant la structure User de l'ami et une variable accept"
        required: true
        content:
          json:
            schema:
              type: object
              properties:
                friend:
                  $ref: "#/components/schemas/User"
                accept:
                  type: boolean
      responses:
        "200":
          description: "Opération réussie"
          content:
            json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    example: "Jean Dupont has been added to your friends."
                  status:
                    type: integer
                    example: 200
        "403":
          description: "Erreur"
...